Use bigdecimal instead of double for rounding

+++ ARPropertyPriceCalculator/app/src/main/java/com/amfam/wombat/arpropertypricecalculator/ARProperty.java
@@ -1,9 +1,9 @@
 package com.amfam.wombat.arpropertypricecalculator;
 
-import java.io.File;
+import java.math.BigDecimal;
 
 public class ARProperty {
     String fileLocation;
-    Double cost;
+    BigDecimal cost;
     String propertyDescription;
 }
